separate class for VerticesNotInMainComponent
